requirements.txt
# Core GCP
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.41.0
google-cloud-monitoring==2.15.1

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Fake Data
faker==25.0.0

# ML
scikit-learn==1.3.2
joblib==1.3.2

# Airflow
apache-airflow==2.7.0
apache-airflow-providers-google==10.1.0

# Web Framework
flask==2.2.5
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utils
python-dotenv==1.0.0
pydantic==1.10.13
